+++ src/routes/blog/+page.svelte
@@ -4,23 +4,31 @@
 	export let data: PageData;
 
 	// Blog Utils
-	import { blogDateFormatter } from './blog-utils';
+	import { getBlogList, blogDateFormatter } from './blog-service';
+
+	function onPrevPage(): void {
+		getBlogList(data.meta.pagination.page - 1).then((res) => (data = res));
+	}
+
+	function onNextPage(): void {
+		getBlogList(data.meta.pagination.page + 1).then((res) => (data = res));
+	}
 </script>
 
 <div class="page-container">
-	<header class="text-center space-y-4">
-		<p class="font-bold">All the latest Skeleton information, straight from the source.</p>
-	</header>
+	<!-- Blog List -->
 	<section class="blog-list space-y-8">
 		{#each data.posts as post, i}
 			<a class="unstyled block hover:card card-body rounded-container-token" href="/blog/{post.slug}" data-sveltekit-prefetch>
 				<article class="grid grid-cols-1 lg:grid-cols-[auto_1fr] gap-4 md:gap-8">
+					<!-- Featured Image -->
 					{#if post.feature_image}
 						<div
 							class="bg-black/50 w-full md:w-64 aspect-square rounded-container-token shadow-xl bg-cover bg-center"
 							style:background-image={`url(${post.feature_image})`}
 						/>
 					{/if}
+					<!-- Content -->
 					<div class="space-y-4">
 						<time class="block">{blogDateFormatter(post.created_at)}</time>
 						<h2>{post.title}</h2>
@@ -36,4 +44,15 @@
 			</a>
 		{/each}
 	</section>
+	<hr />
+	<!-- Pagination -->
+	<footer class="flex justify-between items-center space-x-4">
+		<div>
+			<small class="opacity-50">Page {data.meta.pagination.page} of {data.meta.pagination.pages}</small>
+		</div>
+		<div class="flex items-center space-x-4">
+			<button class="btn-icon btn-filled" on:click={onPrevPage} disabled={!data.meta.pagination.prev}>&larr;</button>
+			<button class="btn btn-filled" on:click={onNextPage} disabled={!data.meta.pagination.next}>Next &rarr;</button>
+		</div>
+	</footer>
 </div>
