--- src/routes/blog/blog-utils.ts
@@ -1,5 +1,0 @@
-export function blogDateFormatter(date: string): string {
-	const options: any = { weekday: 'long', year: 'numeric', month: 'long', day: 'numeric' };
-	const d: Date = new Date(date);
-	return d.toLocaleDateString('en-US', options);
-}
